Added a file that contains all the functions
